+++ tests/profiling/test_scheduler.py
@@ -1,6 +1,9 @@
 # -*- encoding: utf-8 -*-
 import logging
 
+import mock
+
+from ddtrace.internal import compat
 from ddtrace.profiling import event
 from ddtrace.profiling import exporter
 from ddtrace.profiling import recorder
@@ -54,3 +57,13 @@ def test_before_flush_failure(caplog):
     assert caplog.record_tuples == [
         (("ddtrace.profiling.scheduler", logging.ERROR, "Scheduler before_flush hook failed"))
     ]
+
+
+@mock.patch("ddtrace.profiling.scheduler.Scheduler.periodic")
+def test_serverless_periodic(mock_periodic):
+    r = recorder.Recorder()
+    s = scheduler.ServerlessScheduler(r, [exporter.NullExporter()])
+    s._last_export = compat.time_ns() - 65 * 1e9
+    s._profiled_intervals = 65
+    s.periodic()
+    mock_periodic.assert_called()
